Fetch test deps in build_environment.sh.
I discussed this with DeHamer and although the test deps are technically
not required for the build step, in the interest of keeping this
not-really-shared-but-shared-in-spirit file identical across the two and
to not add more bash shenanigans we're just going to download the test
dependencies in the build_environment script no matter what your
intention is. It's just going to cost a tiny bit of extra time on CI for
people who aren't using Godeps.