Remove a few build flags that I'm unsure about.
My reading about cross-compilation in Go tells me that it's using CGO whether
you like it or not. Allowing it to be disabled just produces binaries that
don't work on other platforms. I'm unsure about the installsuffix and ldflags,
and so I'm removing them to not give somebody a false sense that they might
work. Maybe they do, and if that person has a way to test their working or not,
they can add it back. Sorry theoretical future person.